+++ allocators/4lvl-nb/nballoc.h
@@ -60,7 +60,7 @@ typedef struct _taken_list{
 
 
 extern __thread unsigned int myid;
-extern unsigned int number_of_leaves;
+extern unsigned long long number_of_leaves;
 
 void  bd_xx_free(void* n);
 void* bd_xx_malloc(size_t pages);
